+++ bciers/apps/reporting/src/app/bceidbusiness/industry_user_admin/reports/[version_id]/facilities/page.tsx
@@ -1,0 +1,4 @@
+import defaultPageFactory from "@bciers/components/nextPageFactory/defaultPageFactory";
+import Page from "@reporting/src/app/components/reportInformation/facilities/FacilitiesPage";
+
+export default defaultPageFactory(Page);
