+++ bciers/apps/reporting/src/app/components/reportInformation/facilities/fetchFacilitiesPageData.ts
@@ -1,8 +1,8 @@
 import { actionHandler } from "@bciers/actions";
-import { FacilityReportSearchParams } from "@reporting/src/app/components/reportInformation/Facilities/types";
+import { FacilityReportSearchParams } from "@reporting/src/app/components/reportInformation/facilities/types";
 import buildQueryParams from "@bciers/utils/src/buildQueryParams";
 
-// 🛠️ Function to fetch operations
+// 🛠️ Function to fetch facilities
 export const fetchFacilitiesPageData = async (params: {
   version_id: number;
   searchParams: FacilityReportSearchParams;
@@ -10,7 +10,6 @@ export const fetchFacilitiesPageData = async (params: {
   const queryParams = buildQueryParams(params.searchParams);
   const url = `reporting/report-version/${params.version_id}/facility-report-list${queryParams}`;
   const pageData = await actionHandler(url, "GET", url);
-
   return {
     rows: pageData.items,
     row_count: pageData.count,
