+++ bciers/apps/reporting/src/app/components/reportInformation/facilities/types.ts
@@ -3,3 +3,11 @@ export interface FacilityReportSearchParams {
   facility_name?: string;
   facility_bcghgid?: string;
 }
+
+export interface FacilityRow {
+  id: number;
+  facility_id: number;
+  facility_bcghgid: string;
+  facility_name: string;
+  is_completed: boolean;
+}
